pyproject.toml
[tool.poetry]
name = "rezorpay-backend"
version = "0.1.0"
description = ""
authors = ["admin1282 <gaikwadvinayak1282@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
django-split-settings = "^1.3.1"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.2"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"
razorpay = "^1.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"